requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.5
django-appconf==1.0.5
django-channels==0.7.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
idna==3.4
MarkupSafe==2.1.1
oauthlib==3.2.2
pilkit==2.0
Pillow==9.4.0
py-vapid==1.9.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.2
